Code/pyproject.toml
[project]
name = "Optimus-GPT"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "langchain (>=0.3.27,<0.4.0)",
    "matplotlib (>=3.10.7,<4.0.0)",
    "langchain-core (>=0.3.79,<0.4.0)",
    "langchain-openai (>=0.3.35,<0.4.0)",
    "gurobipy (>=12.0.3,<13.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "networkx (>=3.5,<4.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "pydantic (>=2.12.2,<3.0.0)"
    
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
